Revert ".zuul: Run tests only for relevant files"
On a couple of occasions the relevant tests didn't get triggered because
some files weren't listed [1], and on another a commit forgot to update
the list of files [2].

The objective of the CI is to reduce stress for the maintainers, and
make it easy for contributors to find out if their changes work or not.
Missing tests don't help with that, and there's no need to optimize the
tests like this unless there's a real problem to be solved.

[1] Commit deca452b273378e5
    Commit 5c27d730212312d6

[2] Commit b1743c492708826c

This reverts commit c28d902089467884a31de276ec9c8dbb0ef76fa8.

https://github.com/containers/toolbox/pull/1168