cmd/run: Ensure underlying container is stopped when toolbox is killed
Right now "toolbox enter" creates a container on the fly, but
then lets it linger after the foreground toolbox process is
killed (for instance, from a terminal hangup).

Not killing the underlying container has the negative side effect of
stalling shutdown if a toolbox shell is running.

This commit addresses that problem by detecting when the toolbox process
is signaled, and then in response, kills off the entire cgroup associated
with the underlying container.

Closes https://github.com/containers/toolbox/issues/1157

Signed-off-by: Ray Strode <rstrode@redhat.com>