If podman is not needed do not check its config
to be able to build on OpenSUSE.

Following up on fc5f568, which added a check to not require migration if
podman will never be run. Move the check earlier to also not check for the
subuid config.

Note: I see no DCO or definition of what Signed-off-by means in this
project and the CI requires it, so I just sign off on this commit
without it meaning anything.

Signed-off-by: Jan Zerebecki <jan.suse@zerebecki.de>